requirements.txt
# ----------------------
# Flask
# ----------------------
flask
flask-SQLAlchemy
psycopg2
Flask-WTF
flask-marshmallow
# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn
